Initial commit
This was not originally gonna be a git repo, but I later thought
it would be a nice addition.

This repository uses libcsp to
calculate the CRC code for an arbitrary CSP packet given its
header and data segments.

It depends on the `libcsp` repo (found at
https://github.com/sf314/libcsp), which has been added as a git
submodule. Currently, it works as expected when you are on the
`test_enableHeaderCrc` branch of `libcsp`.